Move sparkey side input to scio-extra.
